// src/app/api/model/journalEntryCommand.ts
/**
 * Apache Fineract REST API
 *
 * Contact: dev@fineract.apache.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SingleDebitOrCreditEntryCommand } from './singleDebitOrCreditEntryCommand';


export interface JournalEntryCommand { 
    accountNumber?: string;
    accountingRuleId?: number;
    amount?: number;
    bankNumber?: string;
    checkNumber?: string;
    comments?: string;
    credits?: Array<SingleDebitOrCreditEntryCommand>;
    currencyCode?: string;
    dateFormat?: string;
    debits?: Array<SingleDebitOrCreditEntryCommand>;
    externalAssetOwner?: string;
    locale?: string;
    officeId?: number;
    paymentTypeId?: number;
    receiptNumber?: string;
    referenceNumber?: string;
    routingCode?: string;
    transactionDate?: string;
}